+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/auth/controller/AuthController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ssafy.shinhanflow.auth.service.RefreshTokenService;
+import com.ssafy.shinhanflow.auth.service.AuthService;
 import com.ssafy.shinhanflow.config.error.SuccessResponse;
 
 import lombok.RequiredArgsConstructor;
@@ -16,7 +16,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/auth")
 public class AuthController {
-	private final RefreshTokenService refreshTokenService;
+	private final AuthService refreshTokenService;
 
 	@GetMapping("/access-token")
 	public SuccessResponse<String> refreshAccessToken(@RequestHeader("refresh") String refreshToken) {
