+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/controller/member/MemberController.java
@@ -1,0 +1,24 @@
+package com.ssafy.shinhanflow.controller.member;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.ssafy.shinhanflow.config.error.SuccessResponse;
+import com.ssafy.shinhanflow.dto.member.SignUpRequestDto;
+import com.ssafy.shinhanflow.service.member.MemberService;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequestMapping("/api/v1/members")
+@RequiredArgsConstructor
+public class MemberController {
+	private final MemberService memberService;
+
+	@PostMapping
+	public SuccessResponse<Boolean> createMember(@RequestBody SignUpRequestDto signUpRequestDto) {
+		return SuccessResponse.of(memberService.createMember(signUpRequestDto));
+	}
+}
