+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/dto/FcmMessageRequestDto.java
@@ -1,0 +1,7 @@
+package com.ssafy.shinhanflow.dto;
+
+import lombok.Builder;
+
+@Builder
+public record FcmMessageRequestDto(Long memberId, String title, String content) {
+}
