+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/dto/finance/FinanceApiRequestDto.java
@@ -1,0 +1,7 @@
+package com.ssafy.shinhanflow.dto.finance;
+
+import lombok.Data;
+
+@Data
+public class FinanceApiRequestDto {
+}
