+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/dto/finance/FinanceApiResponseDto.java
@@ -1,0 +1,5 @@
+package com.ssafy.shinhanflow.dto.finance;
+
+// 객체를 직접 생성할 일이 없으므로 추상 클래스로 선언
+public abstract class FinanceApiResponseDto {
+}
