+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/dto/finance/header/ResponseHeaderDto.java
@@ -1,0 +1,16 @@
+package com.ssafy.shinhanflow.dto.finance.header;
+
+import lombok.Data;
+
+@Data
+public class ResponseHeaderDto {
+	private String responseCode;
+	private String responseMessage;
+	private String apiName;
+	private String transmissionDate;
+	private String transmissionTime;
+	private String institutionCode;
+	private String apiKey;
+	private String apiServiceCode;
+	private String institutionTransactionUniqueNo;
+}
