+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/dto/member/SignUpRequestDto.java
@@ -1,0 +1,4 @@
+package com.ssafy.shinhanflow.dto.member;
+
+public record SignUpRequestDto(String email, String password, String name) {
+}
