--- be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/finance/dto/FinanceApiRequestDto.java
@@ -1,7 +1,0 @@
-package com.ssafy.shinhanflow.finance.dto;
-
-import lombok.Data;
-
-@Data
-public class FinanceApiRequestDto {
-}
