--- be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/finance/dto/FinanceApiResponseDto.java
@@ -1,5 +1,0 @@
-package com.ssafy.shinhanflow.finance.dto;
-
-// 객체를 직접 생성할 일이 없으므로 추상 클래스로 선언
-public abstract class FinanceApiResponseDto {
-}
