--- be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/finance/dto/MemberRequestDto.java
@@ -1,11 +1,0 @@
-package com.ssafy.shinhanflow.finance.dto;
-
-import lombok.EqualsAndHashCode;
-import lombok.Value;
-
-@Value
-@EqualsAndHashCode(callSuper = true)
-public class MemberRequestDto extends FinanceApiRequestDto {
-	String apiKey;
-	String userId;
-}
