--- be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/finance/dto/MemberResponseDto.java
@@ -1,13 +1,0 @@
-package com.ssafy.shinhanflow.finance.dto;
-
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
-import lombok.Value;
-
-@Value
-@EqualsAndHashCode(callSuper = true)
-@ToString(callSuper = true)
-public class MemberResponseDto extends FinanceApiResponseDto {
-	String userId; // email
-	String userKey; // pk
-}
