--- be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/finance/dto/account/DemandDepositBalanceResponseDto.java
@@ -1,17 +1,0 @@
-package com.ssafy.shinhanflow.finance.dto.account;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.ssafy.shinhanflow.finance.dto.FinanceApiResponseDto;
-import com.ssafy.shinhanflow.finance.dto.header.ResponseHeaderDto;
-
-public class DemandDepositBalanceResponseDto extends FinanceApiResponseDto {
-	@JsonProperty("Header")
-	ResponseHeaderDto header;
-	@JsonProperty("REC")
-	Rec rec;
-
-	private record Rec(String bankCode, String accountNo, String accountBalance, String accountCreatedDate,
-					   String accountExpiryDate, String lastTransactionDate, String currency) {
-
-	}
-}
