--- be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/finance/dto/header/ResponseHeaderDto.java
@@ -1,16 +1,0 @@
-package com.ssafy.shinhanflow.finance.dto.header;
-
-import lombok.Data;
-
-@Data
-public class ResponseHeaderDto {
-	private String responseCode;
-	private String responseMessage;
-	private String apiName;
-	private String transmissionDate;
-	private String transmissionTime;
-	private String institutionCode;
-	private String apiKey;
-	private String apiServiceCode;
-	private String institutionTransactionUniqueNo;
-}
