--- be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/member/dto/SignUpRequestDto.java
@@ -1,4 +1,0 @@
-package com.ssafy.shinhanflow.member.dto;
-
-public record SignUpRequestDto(String email, String password, String name, String phone) {
-}
