+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/repository/ActionRepository.java
@@ -1,0 +1,8 @@
+package com.ssafy.shinhanflow.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.ssafy.shinhanflow.entity.ActionEntity;
+
+public interface ActionRepository extends JpaRepository<ActionEntity, Long> {
+}
