+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/repository/FlowRepository.java
@@ -1,0 +1,8 @@
+package com.ssafy.shinhanflow.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.ssafy.shinhanflow.entity.FlowEntity;
+
+public interface FlowRepository extends JpaRepository<FlowEntity, Long> {
+}
