+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/repository/MemberRepository.java
@@ -1,0 +1,13 @@
+package com.ssafy.shinhanflow.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.ssafy.shinhanflow.entity.MemberEntity;
+
+@Repository
+public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
+
+	MemberEntity findByEmail(String email);
+
+}
