--- be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/util/FcmMessageRequestDto.java
@@ -1,7 +1,0 @@
-package com.ssafy.shinhanflow.util;
-
-import lombok.Builder;
-
-@Builder
-public record FcmMessageRequestDto(Long memberId, String title, String content) {
-}
